Additional Class Code
I added the Effect and EnemyMonster classes.

Effect is for handling all logic for everything that happens to a monster at the end of a turn. Like DamageOverTime(DOT).

EnemyMonster handles all of the basic enemies that the player will fight.

All attacks must still be added and all effect actions (like being damaged by a poison) must still be added.

I also created various Enumerated variables to help keep track of the different monsters in the game.

I also added an Enum for the Effect names, there will be more effects added later. There functionality must be added as well.

I added a vector and related methods to the Monster class to handle all active effects that a monster has during a fight. This vector will be cleared after every battle.

Also, at the beginning of a fight, the stats of the player monster should be saved. At the end of the fight, those saved stats should be reapplied to the player monster, so that their stats are not permanently affected after a fight.

I believe that the final class to be added next is the BossMonster.

After that final class, we can begin working on the GUI and the actual gameplay. Which should be relatively simple. Then the game will be finished and we can work on the documentation...