Added all ability text and random enemies
I added a function to return a random enemy or the final boss.

I also made it to where the buttons now show the ability text for each player monster.

This is done with several new methods in GUI.

At this point, you can choose a monster, and an enemy will be created, however you cannot use any abilities yet.

ALSO, I added break statements to most switch statements, as they were absent before. The switch in updateAbilityButtons() is intentional.