GAME IS FULLY PLAYABLE
Main game is completed.

Now, polish work is needed.

Need to log every turn. When an ability is used, the textArea is cleared.

So, over the course of a turn we need to tell the player what is happening. What abilities were used, who dodged, who crit, etc...

Also, feel free to change the values of any monster's stats and abilities. It is not balanced at all currently.